% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtex.R
\name{available_tissues}
\alias{available_tissues}
\title{Returns list of tissues in GTEX available for analysis.}
\usage{
available_tissues(expression_matrix_tar = expression_matrix_tar)
}
\arguments{
\item{expression_matrix_tar}{Path to tar file containing expression matricies by tissue.}
}
\value{
Data.frame: polygenic risk score calculated in GTEX samples.
}
\description{
Returns list of tissues in GTEX available for analysis.
}
