% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_prs.R
\name{distanceByPRS}
\alias{distanceByPRS}
\title{Plot distances between transcript profiles}
\usage{
distanceByPRS(all, collapse = F)
}
\arguments{
\item{collapse}{If true all tissues x traits are unlisted and plotted together}

\item{x}{List object returned by "run_eQTL".}
}
\value{
List of plots
}
\description{
Plot distances between transcript profiles
}
