% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_prs.R
\name{distanceByPRS2}
\alias{distanceByPRS2}
\title{Plot distances between transcript profiles from unlisted}
\usage{
distanceByPRS2(l, collapse, plot.eig, tissue = NULL)
}
\value{
List of plots
}
\description{
Plot distances between transcript profiles from unlisted
}
