% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prs.R
\name{extract_query_snps_gtex}
\alias{extract_query_snps_gtex}
\title{harmonise GTEx and query data and calculate polygenic risk score}
\usage{
extract_query_snps_gtex(rsids, gtex_vcf_dir)
}
\arguments{
\item{rsids}{Vector of dbSNP IDs.}

\item{gtex_vcf_dif}{Path to GTEx VCF file.}
}
\value{
Data.frame: polygenic risk score calculated in GTEX samples.
}
\description{
harmonise GTEx and query data and calculate polygenic risk score
}
