% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{extract_top_hits}
\alias{extract_top_hits}
\title{Returns transcripts surpassing given FDR threshold in each tissue}
\usage{
extract_top_hits(x, fdr)
}
\arguments{
\item{x}{List object returned by "run_eQTL"}

\item{fdr}{FDR threshold}
}
\value{
Data.frame of top hits
}
\description{
Returns transcripts surpassing given FDR threshold in each tissue
}
