% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_prs.R
\name{generateSetListGO}
\alias{generateSetListGO}
\title{Generate GO object}
\usage{
generateSetListGO(GOTERMS)
}
\arguments{
\item{GOTERMS}{Vector of GO Terms (GO:#######).}
}
\value{
List Ensembl gene IDS by GO Term
}
\description{
Generate GO object
}
