% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prs.R
\name{get_query_snps_mrbase}
\alias{get_query_snps_mrbase}
\title{generate set of "query" snps using MRBase}
\usage{
get_query_snps_mrbase(outcomes, p)
}
\arguments{
\item{outcomes}{MRBase outcome index.}

\item{p}{Pvalue threshold on which to select SNPs.}
}
\value{
Data.frame of SNPs, betas and alleles.
}
\description{
generate set of "query" snps using MRBase
}
