% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_prs.R
\name{runPathwayAnalysis}
\alias{runPathwayAnalysis}
\title{Run Pathways Analysis}
\usage{
runPathwayAnalysis(x, setList)
}
\arguments{
\item{x}{List object returned by "run_eQTL".}

\item{setList}{Generated using included function}
}
\value{
List pathways scores
}
\description{
Run Pathways Analysis
}
