% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{volcanoeplot}
\alias{volcanoeplot}
\title{Render volcanoe plot for each element in output list.}
\usage{
volcanoeplot(x, fdr, outdir = "../../output/")
}
\arguments{
\item{x}{List object returned by "run_eQTL"}
}
\value{
volcanoe plots
}
\description{
Render volcanoe plot for each element in output list.
}
