% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prs.R
\name{calculate_prs_gtex}
\alias{calculate_prs_gtex}
\title{harmonise GTEx and query data and calculate polygenic risk score}
\usage{
calculate_prs_gtex(query, gtex)
}
\arguments{
\item{query}{Data.frame of "query" SNPs.}

\item{gtex}{Data.frame og GTEx SNPs.}
}
\value{
Data.frame: polygenic risk score calculated in GTEX samples.
}
\description{
harmonise GTEx and query data and calculate polygenic risk score
}
