% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prs.R
\name{calculate_prs_gtex}
\alias{calculate_prs_gtex}
\title{harmonise GTEx and query data and calculate polygenic risk score}
\usage{
calculate_prs_gtex(query_data_frame, gtex_data_frame)
}
\arguments{
\item{query_data_frame}{Data.frame with four columns: 'SNP','effect_allele','other_allele','beta'.}

\item{gtex_data_frame}{Returned by extract_query_snps_gtex function}
}
\value{
Data.frame: polygenic risk score calculated in GTEX samples.
}
\description{
harmonise GTEx and query data and calculate polygenic risk score
}
