% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prs.R
\name{extract_query_snps_gtex}
\alias{extract_query_snps_gtex}
\title{Given a set of dbSNP rsids, get genome coordinates (hg19) and extract these ranges from GTEx VCF}
\usage{
extract_query_snps_gtex(rsids, gtex_vcf_dir)
}
\arguments{
\item{rsids}{Vector of SNPs in dbSNP format.}

\item{gtex_vcf_dif}{Path to GTEx VCF file.}
}
\value{
Data.frame: GTEx alleles matching query.
}
\description{
Given a set of dbSNP rsids, get genome coordinates (hg19) and extract these ranges from GTEx VCF
}
