% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{extract_top_hits}
\alias{extract_top_hits}
\title{Returns transcripts surpassing given FDR threshold in each tissue}
\usage{
extract_top_hits(x, fdr = T, pthresh = 0.05)
}
\arguments{
\item{x}{List object returned by "run_eQTL".}

\item{fdr}{Should FDR be used.}

\item{pthresh}{Threshold for annotation gene names.}
}
\value{
Data.frame of genes passing threshold, annotated with gene names.
}
\description{
Returns transcripts surpassing given FDR threshold in each tissue
}
