% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{volcanoeplot}
\alias{volcanoeplot}
\title{Render volcanoe plot for each element in output list.}
\usage{
volcanoeplot(x, fdr = T, p.thresh = 0.05)
}
\arguments{
\item{x}{List object returned by "run_eQTL"}

\item{fdr}{Should FDR corrected p-values be used for annotation.}

\item{p.thresh}{Threshold for adding gene name label.}
}
\value{
volcanoe plots
}
\description{
Render volcanoe plot for each element in output list.
}
